requirement.txt
asgiref==3.4.1
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cssselect==1.1.0
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-currentuser==0.5.0
django-debug-toolbar==2.2
django-excel-response==2.0.5
django-import-export==2.4.0
django-jsignature==0.11
django-mssql==1.8
django-pdfkit==0.3.1
django-phone-field==1.8.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pytds==1.5
django-six==1.0.4
django-sqlserver==1.11
django-tables2==2.3.4
django-uuslug==1.2.0
django-weasyprint==1.1.0.post1
django-wkhtmltopdf==3.2.0
et-xmlfile==1.0.1
fonttools==4.27.1
greenlet==1.1.0
html5lib==1.1
idna==3.3
jdcal==1.4.1
JPype1 @ file:///D:/webapps/virtualenv/TLC_HRS/hrs/JPype1-1.2.0-cp38-cp38-win32.whl
ldap3==2.7
lxml==4.8.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==2.6.0
pdfkit==0.6.0
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydyf==0.1.1
pymssql==2.1.5
pyodbc==4.0.30
pyphen==0.11.0
pyquery==1.4.3
pyreportjasper==1.0.2
python-ldap @ file:///D:/webapps/virtualenv/TLC_HRS/hrs/python_ldap-3.3.1-cp38-cp38-win32.whl
python-slugify==4.0.1
python-tds==1.10.0
pytz==2020.1
pywin32==300
PyYAML==5.3.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.4.15
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
tinycss2==1.1.0
tlc-component @ file:///D:/webapps/virtualenv/TLC_HRS/tlc_component-1.5.3-py3-none-any.whl
tlc-ldap-auth @ file:///D:/webapps/virtualenv/TLC_HRS/hrs/tlc_ldap_auth-1.3.3-py3-none-any.whl
urllib3==1.26.7
WeasyPrint==52.5
webencodings==0.5.1
wkhtmltopdf==0.2
xlrd==1.2.0
xlwt==1.3.0
zopfli==0.1.8
